Refactor API to clear todo item
