+++ apps/server/src/routers/api/filters/mod.rs
@@ -1,0 +1,9 @@
+mod library;
+mod media;
+mod metadata;
+mod series;
+
+pub(crate) use library::*;
+pub(crate) use media::*;
+pub(crate) use metadata::*;
+pub(crate) use series::*;
