+++ apps/server/src/routers/api/mod.rs
@@ -1,8 +1,9 @@
+pub(crate) mod filters;
+pub(crate) mod v1;
+
 use axum::Router;
 
 use crate::config::state::AppState;
-
-pub(crate) mod v1;
 
 pub(crate) fn mount(app_state: AppState) -> Router<AppState> {
 	Router::new().nest("/api", Router::new().nest("/v1", v1::mount(app_state)))
@@ -17,12 +18,40 @@ mod tests {
 		NamedType,
 	};
 
-	use crate::config::jwt::CreatedToken;
-
-	use super::v1::{
-		auth::*, book_club::*, emailer::*, epub::*, job::*, library::*, media::*,
-		metadata::*, series::*, smart_list::*, user::*, ClaimResponse, StumpVersion,
-		UpdateCheck,
+	use crate::{
+		config::jwt::CreatedToken,
+		routers::api::v1::{
+			auth::{AuthenticationOptions, LoginOrRegisterArgs, LoginResponse},
+			book_club::{
+				BookClubInvitationAnswer, CreateBookClub, CreateBookClubInvitation,
+				CreateBookClubMember, CreateBookClubSchedule, CreateBookClubScheduleBook,
+				CreateBookClubScheduleBookOption, GetBookClubsParams, UpdateBookClub,
+				UpdateBookClubMember,
+			},
+			emailer::{
+				CreateOrUpdateEmailDevice, CreateOrUpdateEmailer, EmailerIncludeParams,
+				EmailerSendRecordIncludeParams, PatchEmailDevice,
+				SendAttachmentEmailResponse, SendAttachmentEmailsPayload,
+			},
+			epub::{CreateOrUpdateBookmark, DeleteBookmark},
+			job::UpdateSchedulerConfig,
+			library::{
+				CleanLibraryResponse, CreateLibrary, LibraryStatsParams,
+				PatchLibraryThumbnail, UpdateLibrary, UpdateLibraryExcludedUsers,
+			},
+			media::{
+				individual::{MediaIsComplete, PutMediaCompletionStatus},
+				thumbnails::PatchMediaThumbnail,
+			},
+			metadata::MediaMetadataOverview,
+			series::{PatchSeriesThumbnail, SeriesIsComplete},
+			smart_list::{
+				CreateOrUpdateSmartList, CreateOrUpdateSmartListView,
+				GetSmartListsParams, SmartListMeta, SmartListRelationOptions,
+			},
+			user::{CreateUser, DeleteUser, UpdateUser, UpdateUserPreferences},
+			ClaimResponse, StumpVersion, UpdateCheck,
+		},
 	};
 
 	#[allow(dead_code)]
@@ -147,9 +176,8 @@ mod tests {
 			format!("{}\n\n", ts_export::<CreateBookClubSchedule>()?).as_bytes(),
 		)?;
 
-		file.write_all(
-			format!("{}\n\n", ts_export::<PatchMediaThumbnail>()?).as_bytes(),
-		)?;
+		let var_name = &format!("{}\n\n", ts_export::<PatchMediaThumbnail>()?);
+		file.write_all(var_name.as_bytes())?;
 		file.write_all(
 			format!("{}\n\n", ts_export::<PatchSeriesThumbnail>()?).as_bytes(),
 		)?;
