+++ apps/server/src/routers/api/v1/mod.rs
@@ -13,23 +13,6 @@ use crate::{
 	errors::{APIError, APIResult},
 };
 
-// TODO: I think this module might benefit from being split up a bit, especially for some of the larger
-// routers like media.rs. It's a bit unwieldy to have everything in one file, perhaps something like:
-// |- media
-// |  |- mod.rs (mounts the router, etc)
-// |  |- bulk.rs (bulk media operations)
-// |  |- (by_)?id.rs (media by ID operations)
-// |  |- utils.rs (misc media operations, filters, etc)
-// OR (actually) wrt utils, maybe have a separate module for utils, e.g. filters, in the api root since
-// most likely they will be shared across different API versions. So maybe something like:
-// |- api
-// |  |- mod.rs (mounts the router, etc)
-// |  |- filter.rs (filters for the api, e.g. apply_media_read_status_filter)
-// |  |- v1/*/etc
-// This isn't really a big deal, but it might make the code a bit easier to navigate and understand
-// for new contributors. Not that there are many contributors lol but still. Maybe I just want it for
-// my own sanity 🙈
-
 // TODO: Also, there is a lot of cringe and smell throughout some of the older code. While I definitely want to focus on building
 // out new features and fixing bugs, I think it would be a good idea to start cleaning up some of the older code when I have time. I
 // also think there is a good amount of duplication which can be trimmed down, like how I did with the OPDS v2 API. A few of those route
