+++ apps/server/src/routers/opds/v1_2.rs
@@ -26,18 +26,15 @@ use crate::{
 	errors::{APIError, APIResult},
 	filter::chain_optional_iter,
 	middleware::auth::{auth_middleware, RequestContext},
-	routers::api::v1::{
-		library::library_not_hidden_from_user_filter,
-		media::{
-			apply_media_library_not_hidden_for_user_filter, get_media_thumbnail_by_id,
+	routers::api::{
+		filters::{
+			apply_in_progress_filter_for_user, apply_media_age_restriction,
+			apply_media_library_not_hidden_for_user_filter, apply_series_age_restriction,
+			library_not_hidden_from_user_filter,
 		},
+		v1::media::thumbnails::get_media_thumbnail_by_id,
 	},
 	utils::http::{ImageResponse, NamedFile, Xml},
-};
-
-use crate::routers::api::v1::{
-	media::{apply_in_progress_filter_for_user, apply_media_age_restriction},
-	series::apply_series_age_restriction,
 };
 
 pub(crate) fn mount(app_state: AppState) -> Router<AppState> {
