+++ apps/server/src/routers/utoipa.rs
@@ -23,8 +23,8 @@ use crate::middleware::auth::auth_middleware;
 use super::api::{
 	self,
 	v1::{
-		auth::LoginOrRegisterArgs, library::*, media::*, notifier::*, series::*,
-		smart_list::*, user::*, ClaimResponse, StumpVersion,
+		auth::LoginOrRegisterArgs, library::*, media::individual::*, notifier::*,
+		series::*, smart_list::*, user::*, ClaimResponse, StumpVersion,
 	},
 };
 
@@ -64,20 +64,20 @@ use super::api::{
         api::v1::library::delete_library,
         api::v1::log::get_logs,
         api::v1::log::delete_logs,
-        api::v1::media::get_media,
-        api::v1::media::get_duplicate_media,
-        api::v1::media::get_in_progress_media,
-        api::v1::media::get_recently_added_media,
-        api::v1::media::get_media_by_id,
-        api::v1::media::get_media_file,
-        api::v1::media::convert_media,
-        api::v1::media::get_media_page,
-        api::v1::media::get_media_thumbnail_handler,
-        api::v1::media::update_media_progress,
-        api::v1::media::get_media_progress,
-        api::v1::media::delete_media_progress,
-        api::v1::media::get_is_media_completed,
-        api::v1::media::put_media_complete_status,
+        api::v1::media::bulk::get_media,
+        api::v1::media::bulk::get_duplicate_media,
+        api::v1::media::bulk::get_in_progress_media,
+        api::v1::media::bulk::get_recently_added_media,
+        api::v1::media::individual::get_media_by_id,
+        api::v1::media::individual::get_media_file,
+        api::v1::media::individual::convert_media,
+        api::v1::media::individual::get_media_page,
+        api::v1::media::thumbnails::get_media_thumbnail_handler,
+        api::v1::media::individual::update_media_progress,
+        api::v1::media::individual::get_media_progress,
+        api::v1::media::individual::delete_media_progress,
+        api::v1::media::individual::get_is_media_completed,
+        api::v1::media::individual::put_media_complete_status,
         api::v1::metadata::get_metadata_overview,
         api::v1::metadata::get_genres_handler,
         api::v1::metadata::get_writers_handler,
